feat: added support for reading a custom tag from cli arguments
