+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@morlay/bunpublish",
-  "version": "0.1.7",
+  "version": "0.2.0",
   "type": "module",
   "publishConfig": {
     "access": "public"
@@ -11,17 +11,19 @@
   "dependencies": {
     "@npmcli/arborist": "^7.2.1",
     "chalk": "^5.3.0",
+    "commander": "^12.0.0",
     "globby": "^14.0.0",
     "libnpmpublish": "^9.0.1",
     "npm-registry-fetch": "^16.1.0",
     "pacote": "^17.0.4"
   },
   "devDependencies": {
-    "bun-types": "^1.0.11",
-    "typescript": "^5.2.2",
-    "prettier": "^3.0.3",
+    "@types/bun": "^1.0.5",
+    "@types/libnpmpublish": "^4.0.6",
     "@types/pacote": "^11.1.8",
-    "@types/libnpmpublish": "^4.0.6"
+    "bun-types": "^1.0.11",
+    "prettier": "^3.0.3",
+    "typescript": "^5.2.2"
   },
   "bin": {
     "bunpublish": "./bin/index.ts"
